chore: Add classcache system to fix tokio find_class issues
On android, threads created by tokio, even when registered with the JVM, don't seem to success when
find_class is called on symbols loaded from native library.
Create a class cache where we can load class defs on the
main thread then cache them for use on other threads
when needed as GlobalRefs.